Move methods into Timeline
